requirements.txt
pandas
numpy
pyyaml
python-dotenv
fastapi
uvicorn
joblib
jupyter
ipykernel
matplotlib
seaborn
scipy
scikit-learn
pyarrow
lightgbm
mlflow
dagshub
xgboost
shap
confluent_kafka
kafka-python
streamlit